Update setup.py and setup.cfg
Summary:
The changes:

setup.cfg now also has a pycodestyle section. See https://github.com/PyCQA/pycodestyle/issues/466 for the reason why (thanks guido!)

setup.py now calls find_packages. This is neccecary if your package folder contains subfolders. for example 'sam.preprocessing' is a seperate 'package' for the purpose of the packages option. So we now get everything instead of just the root dofolder.

setup.py now has a long_description as well. Not sure when this is ever relevant, butwhy not.

Test Plan: have the same with SAM, and it works there

Reviewers: Sebastiaan

Reviewed By: Sebastiaan

Differential Revision: https://dev.ynformed.nl/D615